clean up and make App root
