+++ src/web/components/MainPage/types.ts
@@ -1,0 +1,1 @@
+export type MainPageProps = {};
